requirements.txt
streamlit
spotipy
pillow
python-dotenv
requests